requirements/local_requirement.txt
-r base_requirement.txt

django-extensions
django-debug-toolbar
